+++ TROPIC01_fw_update_files/boot_v_1_0_1/fw_v_2_0_0/fw_SPECT.h
@@ -1,4 +1,7 @@
 #pragma once
+
+#include <stdint.h>
+
 /**
  * @brief SPECT firmware version spect_app-v1.0.0_signed.bin for bootloader v1.0.1
  */