+++ TROPIC01_fw_update_files/boot_v_2_0_1/fw_v_2_0_0/fw_CPU.h
@@ -1,4 +1,7 @@
 #pragma once
+
+#include <stdint.h>
+
 /**
  * @brief CPU firmware version fw_v2.0.0.hex32_signed_chunks.bin for bootloader v2.0.1
  */