+++ tests/common/lt_test_common.c
@@ -16,7 +16,6 @@
 
 #include "libtropic.h"
 #include "libtropic_common.h"
-#include "libtropic_functional_tests.h"
 #include "libtropic_logging.h"
 
 lt_ret_t (*lt_test_cleanup_function)(void) = NULL;