+++ tests/functional_mock/lt_test_mock_hardware_fail.c
@@ -12,7 +12,6 @@
 
 #include "libtropic.h"
 #include "libtropic_common.h"
-#include "libtropic_functional_tests.h"
 #include "libtropic_l2.h"
 #include "libtropic_logging.h"
 #include "libtropic_macros.h"